The wheel automatically returns to default rotation when not steered.
Whenever the wheel is not beeing steered, it will put itself back into an upright position.
When we code the steering of the truck, instead of contimuously listening to the wheel's rotation, we need to only adjust the truck's horizontal movement while the wheel is actually held.  The Wheel has a public GetSteered method, which returns true on frames when the wheel is actively steered by a hand. At all other times, the truck should not move horizontally.